+++ lighthouse-cli/test/smokehouse/test-definitions/pwa-caltrain.js
@@ -5,60 +5,13 @@
  */
 'use strict';
 
-/** @fileoverview Expected Lighthouse audit values for various sites with stable(ish) PWA results. */
-
 import pwaDetailsExpectations from './pwa-expectations-details.js';
 
-const jakeExpectations = {...pwaDetailsExpectations, hasShortName: false};
-
-/**
- * @type {Smokehouse.ExpectedRunnerResult}
- * Expected Lighthouse results for svgomg.
- */
-const svgomg = {
-  lhr: {
-    requestedUrl: 'https://jakearchibald.github.io/svgomg/',
-    finalUrl: 'https://jakearchibald.github.io/svgomg/',
-    audits: {
-      'service-worker': {
-        score: 1,
-      },
-      'viewport': {
-        score: 1,
-      },
-      'installable-manifest': {
-        score: 1,
-        details: {items: [], debugData: {manifestUrl: 'https://jakearchibald.github.io/svgomg/manifest.json'}},
-      },
-      'splash-screen': {
-        score: 1,
-        details: {items: [jakeExpectations]},
-      },
-      'themed-omnibox': {
-        score: 1,
-        details: {items: [jakeExpectations]},
-      },
-      'content-width': {
-        score: 1,
-      },
-      'apple-touch-icon': {
-        score: 1,
-      },
-
-      // "manual" audits. Just verify in the results.
-      'pwa-cross-browser': {
-        score: null,
-        scoreDisplayMode: 'manual',
-      },
-      'pwa-page-transitions': {
-        score: null,
-        scoreDisplayMode: 'manual',
-      },
-      'pwa-each-page-has-url': {
-        score: null,
-        scoreDisplayMode: 'manual',
-      },
-    },
+/** @type {LH.Config.Json} */
+const config = {
+  extends: 'lighthouse:default',
+  settings: {
+    onlyCategories: ['pwa'],
   },
 };
 
@@ -66,7 +19,7 @@ const svgomg = {
  * @type {Smokehouse.ExpectedRunnerResult}
  * Expected Lighthouse results for caltrainschedule.io.
  */
-const caltrain = {
+const expectations = {
   lhr: {
     requestedUrl: 'https://caltrainschedule.io/',
     finalUrl: 'https://caltrainschedule.io/',
@@ -112,7 +65,8 @@ const caltrain = {
   },
 };
 
-export {
-  svgomg,
-  caltrain,
+export default {
+  id: 'pwa-caltrain',
+  expectations,
+  config,
 };
