+++ lighthouse-cli/test/smokehouse/test-definitions/pwa-expectations-details.js
